Add Crawl Component
This component acts as the crawl display. Currently it just plays the
crawl until timeout. It would be nice to figure out a way to read the
css values directly to figure out how long it takes to play, but I
don't see a clear way to do that for now.