Add "CrawlEntry" component
This basically acts as the wrapper around the Crawl. It takes multiple
components and wraps them together. In theory, after the main crawl is
done I could then have this component do the "Pan down to the planet"
part of Star Wars openings, but for now that isn't a supported
feature.